Доработана StationInfoBottomSheetViewModel.kt - получение адреса по координатам из вью модели.
