+++ profileScreen/src/main/java/chargeit/profilescreen/view/fragent/ProfileFragment.kt
@@ -73,11 +73,9 @@ class ProfileFragment : CoreFragment(R.layout.fragment_profile) {
     private fun initListeners() {
         binding.registrationButton.setOnClickListener {
             viewModel.navigateToRegistrationScreen()
-            //it.findNavController().navigate(R.id.profile_registration_fragment)
         }
         binding.enterButton.setOnClickListener {
             viewModel.navigateToLoginScreen()
-            //it.findNavController().navigate(R.id.login_fragment)
         }
         binding.changeData.setOnClickListener {
             Toast.makeText(requireContext(), "Данные пока нельзя изменить", Toast.LENGTH_SHORT)
